+++ addons/addon-base-raas-ui/packages/base-raas-ui/src/parts/environments-sc/ScEnvironmentsList.js
@@ -17,7 +17,7 @@ import _ from 'lodash';
 import { decorate, computed, action, observable, runInAction } from 'mobx';
 import { observer, inject } from 'mobx-react';
 import { withRouter } from 'react-router-dom';
-import { Container, Segment, Header, Icon, Button, Label } from 'semantic-ui-react';
+import { Container, Segment, Header, Icon, Form, Grid, Input, Dropdown } from 'semantic-ui-react';
 
 import { gotoFn } from '@aws-ee/base-ui/dist/helpers/routing';
 import { swallowError, storage } from '@aws-ee/base-ui/dist/helpers/utils';
@@ -35,6 +35,17 @@ import ProgressPlaceHolder from '@aws-ee/base-ui/dist/parts/helpers/BasicProgres
 import { filterNames } from '../../models/environments-sc/ScEnvironmentsStore';
 import ScEnvironmentCard from './ScEnvironmentCard';
 import ScEnvironmentsFilterButtons from './parts/ScEnvironmentsFilterButtons';
+import EnvsHeader from './ScEnvsHeader';
+
+const envOptions = [
+  { key: 'any', text: 'Any Attribute', value: 'any' },
+  { key: 'meta', text: 'Name or Description', value: 'meta' },
+  { key: 'user', text: 'User', value: 'user' },
+  { key: 'project', text: 'Project', value: 'project' },
+  { key: 'type', text: 'Workspace Type', value: 'type' },
+  { key: 'config-type', text: 'Configuration Name', value: 'configType' },
+  { key: 'study', text: 'Study', value: 'study' },
+];
 
 // expected props
 // - scEnvironmentsStore (via injection)
@@ -49,6 +60,8 @@ class ScEnvironmentsList extends React.Component {
       storage.setItem(key, name);
       this.selectedFilter = name;
       this.provisionDisabled = false;
+      this.searchType = 'any';
+      this.search = '';
     });
   }
 
@@ -81,6 +94,14 @@ class ScEnvironmentsList extends React.Component {
     return this.props.scEnvironmentsStore;
   }
 
+  get viewStore() {
+    return this.props.ScEnvView;
+  }
+
+  get isAdmin() {
+    return this.props.userStore.user.isAdmin;
+  }
+
   getProjects() {
     const store = this.getProjectsStore();
     return store.list;
@@ -106,9 +127,14 @@ class ScEnvironmentsList extends React.Component {
     storage.setItem(key, name);
   };
 
+  handleViewToggle() {
+    return () => runInAction(() => this.viewStore.toggleView());
+  }
+
   render() {
     const store = this.envsStore;
     let content = null;
+    let list = [];
     const projects = this.getProjects();
     const appStreamProjectIds = _.map(
       _.filter(projects, proj => proj.isAppStreamConfigured),
@@ -126,14 +152,24 @@ class ScEnvironmentsList extends React.Component {
     } else if (isStoreEmpty(store)) {
       content = this.renderEmpty();
     } else if (isStoreNotEmpty(store)) {
-      content = this.renderMain();
-    } else {
-      content = null;
+      list = this.searchAndFilter();
+      content = (
+        <>
+          <EnvsHeader
+            current={list.length}
+            total={store.total}
+            isAdmin={this.isAdmin}
+            provisionDisabled={this.provisionDisabled}
+            onViewToggle={this.handleViewToggle()}
+            onEnvCreate={this.handleCreateEnvironment}
+          />
+          {this.renderMain(list)}
+        </>
+      );
     }
 
     return (
       <Container className="mt3 animated fadeIn">
-        {this.renderTitle()}
         {this.provisionDisabled && this.renderMissingAppStreamConfig()}
         {content}
       </Container>
@@ -157,19 +193,88 @@ class ScEnvironmentsList extends React.Component {
     );
   }
 
-  renderMain() {
-    const store = this.envsStore;
-    const selectedFilter = this.selectedFilter;
-    const list = store.filtered(selectedFilter);
+  searchAndFilter() {
+    const list = this.envsStore.filtered(this.selectedFilter);
+    if (_.isEmpty(this.search)) return list;
+    const searchString = `(${_.escapeRegExp(this.search).replace(' or ', '|')})`;
+    const exp = new RegExp(searchString, 'i');
+
+    const userName = env => this.props.userDisplayName.getDisplayName({ uid: env.createdBy });
+    const workspaceName = env => this.props.envTypesStore.getEnvType(env.envTypeId)?.name;
+    const configName = env =>
+      this.props.envTypesStore.getEnvTypeConfigsStore(env.envTypeId).getEnvTypeConfig(env.envTypeConfigId)?.name;
+
+    const searchMap = {
+      any: env =>
+        !_.isEmpty(
+          _.find(
+            [
+              env.name,
+              env.description,
+              env.projectId,
+              env.studyIds.join(', '),
+              userName(env),
+              workspaceName(env),
+              configName(env),
+            ],
+            val => exp.test(val),
+          ),
+        ),
+      meta: env => exp.test(env.name) || exp.test(env.description),
+      user: env => exp.test(userName(env)),
+      project: env => exp.test(env.projectId),
+      type: env => exp.test(workspaceName(env)),
+      configType: env => exp.test(configName(env)),
+      study: env => exp.test(env.studyIds.join(', ')),
+    };
+
+    return _.filter(list, searchMap[this.searchType]);
+  }
+
+  renderMain(list) {
     const isEmpty = _.isEmpty(list);
 
     return (
       <div data-testid="workspaces">
-        <ScEnvironmentsFilterButtons
-          selectedFilter={selectedFilter}
-          onSelectedFilter={this.handleSelectedFilter}
-          className="mb3"
-        />
+        <Form>
+          <Grid columns={2} stackable className="mt2">
+            <Grid.Row stretched>
+              <Grid.Column width={6}>
+                <Input
+                  fluid
+                  action={
+                    <Dropdown
+                      button
+                      basic
+                      floating
+                      defaultValue="any"
+                      options={envOptions}
+                      onChange={(e, data) =>
+                        runInAction(() => {
+                          this.searchType = data.value;
+                        })
+                      }
+                    />
+                  }
+                  placeholder="Search"
+                  icon="search"
+                  iconPosition="left"
+                  onChange={(e, data) =>
+                    runInAction(() => {
+                      this.search = data.value;
+                    })
+                  }
+                />
+              </Grid.Column>
+              <Grid.Column width={10}>
+                <ScEnvironmentsFilterButtons
+                  selectedFilter={this.selectedFilter}
+                  onSelectedFilter={this.handleSelectedFilter}
+                />
+              </Grid.Column>
+            </Grid.Row>
+          </Grid>
+        </Form>
         {!isEmpty &&
           _.map(list, item => (
             <Segment className="p3 mb4" clearing key={item.id}>
@@ -180,7 +285,7 @@ class ScEnvironmentsList extends React.Component {
           <Segment placeholder>
             <Header icon className="color-grey">
               <Icon name="server" />
-              No research workspaces matching the selected filter.
+              No research workspaces matching the selected filter or search.
               <Header.Subheader>Select &apos;All&apos; to view all the workspaces</Header.Subheader>
             </Header>
           </Segment>
@@ -200,36 +305,6 @@ class ScEnvironmentsList extends React.Component {
       </Segment>
     );
   }
-
-  renderTitle() {
-    return (
-      <div className="mb3 flex">
-        <Header as="h3" className="color-grey mt1 mb0 flex-auto">
-          <Icon name="server" className="align-top" />
-          <Header.Content className="left-align">Research Workspaces {this.renderTotal()}</Header.Content>
-        </Header>
-        <div>
-          <Button
-            data-testid="create-workspace"
-            color="blue"
-            size="medium"
-            disabled={this.provisionDisabled}
-            basic
-            onClick={this.handleCreateEnvironment}
-          >
-            Create Research Workspace
-          </Button>
-        </div>
-      </div>
-    );
-  }
-
-  renderTotal() {
-    const store = this.envsStore;
-    if (isStoreError(store) || isStoreLoading(store)) return null;
-
-    return <Label circular>{store.total}</Label>;
-  }
 }
 
 // see https://medium.com/@mweststrate/mobx-4-better-simpler-faster-smaller-c1fbc08008da
@@ -238,12 +313,20 @@ decorate(ScEnvironmentsList, {
   provisionDisabled: observable,
   envsStore: computed,
   envTypesStore: computed,
+  viewStore: computed,
+  isAdmin: computed,
   handleCreateEnvironment: action,
   handleSelectedFilter: action,
+  handleViewToggle: action,
+  search: observable,
 });
 
 export default inject(
+  'ScEnvView',
   'scEnvironmentsStore',
   'projectsStore',
   'envTypesStore',
+  'userDisplayName',
+  'envTypesStore',
+  'userStore',
 )(withRouter(observer(ScEnvironmentsList)));